codeql UB: pointer overflow check
When checking for integer overflow, you may often write tests like p + i <
p.  This works fine if p and i are unsigned integers, since any overflow in
the addition will cause the value to simply "wrap around." However, using
this pattern when p is a pointer is problematic because pointer overflow has
undefined behavior according to the C and C++ standards.  If the addition
overflows and has an undefined result, the comparison will likewise be
undefined; it may produce an unintended result, or may be deleted entirely
by an optimizing compiler.

fixes codeql #51, #52

+++ src/decode.c
@@ -4803,9 +4803,9 @@ dwg_decode_add_object (Dwg_Data *restrict dwg, Bit_Chain *dat,
   bit_reset_chain (dat);
   if (obj->size > dat->size ||
       dat->size > abs_dat.size ||
-      dat->chain + dat->size < dat->chain ||
-      abs_dat.chain + abs_dat.size < abs_dat.chain ||
-      dat->chain + dat->size > abs_dat.chain + abs_dat.size)
+      (long)(dat->chain + dat->size) < (long)dat->chain ||
+      (long)(abs_dat.chain + abs_dat.size) < (long)abs_dat.chain ||
+      (long)(dat->chain + dat->size) > (long)(abs_dat.chain + abs_dat.size))
     {
       LOG_TRACE ("\n");
       LOG_WARN ("Invalid object size %u > %ld. Would overflow", obj->size,
